Refactor Dhis2Tei module
