+++ src/webapp/logic/dhisConnector.js
@@ -1,5 +1,6 @@
 import _ from "lodash";
 import { promiseMap } from "../utils/promises";
+import { getTrackerProgramMetadata } from "../../data/Dhis2RelationshipTypes";
 
 export async function getElement(api, type, id) {
     const endpoint = type === "dataSets" ? "dataSets" : "programs";
@@ -45,12 +46,7 @@ export async function getElementMetadata({ element, api, orgUnitIds }) {
             .getData()
     );
 
-    const metadata = await api.metadata
-        .get({
-            relationshipTypes: { fields: { id: true, name: true } },
-        })
-        .getData();
-
+    const metadata = element.type === "trackerPrograms" ? await getTrackerProgramMetadata(api) : {};
     const organisationUnits = _.flatMap(responses, ({ organisationUnits }) => organisationUnits);
 
     return { element, metadata, elementMetadata, organisationUnits, rawMetadata };
